### 2023-02-26 #5188 @editorialbot
[PRE REVIEW]: fRAT: an interactive, Python-based tool for region-of-interest summaries of functional imaging data
**Submitting author:** @elliohow (<a href="http://orcid.org/0000-0002-3868-2516">Elliot Howley</a>)
**Repository:** https://github.com/elliohow/fMRI_ROI_Analysis_Tool
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
