### 2023-02-27 #5188 @Kevin-Mattheus-Moerman
@elliohow I am the AEiC on this track and here to help with initial steps. Below are some comments that need your attention.

- [ ] In your paper, for the affiliation, please spell out United Kingdom, and add Nottingham as the town (you may also consider adding a discipline/department within the university if you like, to make it a bit more complete).
- [ ] You may replace `open-source` with `open source` in the paper.
- [ ] On your repository, GitHub does not seem to recognise your license text as Apache 2.0. I had a look and I was wondering if it is because of the added text `Prior to release 1.0.0, fRAT was licensed under a MIT license`. Please remove this text (I would add a License section to the readme and mention the current and this past license there), and ensure that your open source license text fully matches (apart from your own name being added of course) with the official text https://opensource.org/license/apache-2-0/.
