+++ README.md
@@ -35,4 +35,5 @@ We use [Semantic versioning](http://semver.org/) for versioning. For the version
 
 ## Licensing
 This project uses the Apache 2.0 license. For the text version of the license see
-[here](https://github.com/elliohow/fMRI_ROI_Analysis_Tool/blob/master/LICENSE).
+[here](https://github.com/elliohow/fMRI_ROI_Analysis_Tool/blob/master/LICENSE). 
+Prior to version 1.0.0, this project used an MIT license.
