### 2023-02-28 #5188 @mstimberg
:wave: Thanks for submitting your paper to JOSS @elliohow. I will oversee the reviewing process.

If you have any suggestions for potential reviewers, then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
