### 2023-02-28 #5188 @ZeitgeberH
@mstimberg  Thanks for the invitation.
I checked the doc of the submission. It seems target to MacOS. Unfortunately, I do not have an activate MacOS right now. Since  FSL (that fRAT relies on) works on WSL2 in Windows10, I can test if it works there (and possibly Ubuntu) if  necessary.
