### 2023-02-28 #5188 @mstimberg
Thanks, @ZeitgeberH, I had not seen this.
@elliohow Can you please comment on this: is macOS necessary to use/test your software, or can it also run on Linux/Windows (e.g. via WSL)?
