### 2023-02-28 #5188 @elliohow
@mstimberg I haven't tested whether it works using either WSL or Linux, but I just saw that FSL and Nipype can still work well together in WSL which I thought may be the main hiccup ([link](https://neurostars.org/t/difficulty-with-fsl-and-nipype-following-fmriprep-preprocessing/19174/5)). But there may be still be some unexpected errors when using WSL or Linux.
