### 2023-02-28 #5188 @billbrod
@mstimberg and @elliohow I have used FSL and Nipype on Linux  (together and separate) without a problem (don't know about WSL). I also do not have access to a MacOS machine, but can review on Linux (Fedora, Ubuntu, or Rocky).
