### 2023-02-28 #5188 @mstimberg
Thanks @billbrod, that's very useful to know.

@elliohow Would you be happy with reviewers that run their tests on Linux (and/or Windows WSL)? It might mean some extra work for you in case they run into OS-dependent issues, but it also would mean you could advertise your software to a bigger target audience if everything goes smoothly.

And thanks again @ZeitgeberH and @billbrod for agreeing to review. I will assign you in case @elliohow agrees to go forward.
