+++ README.md
@@ -21,6 +21,21 @@ Documentation:
 
 [ROI analysis tutorial](https://fmri-roi-analysis-tool.readthedocs.io/en/latest/tutorials/Basic-ROI-analysis.html)
 
+## Reporting bugs
+
+To report a bug, please go to [fRAT's Issues](https://github.com/elliohow/fMRI_ROI_Analysis_Tool/issues/new).
+
+For other questions, issues or discussion please go to [fRAT's Discussions](https://github.com/elliohow/fMRI_ROI_Analysis_Tool/discussions).
+
+## Contributing with development
+
+The [Fork & Pull Request Workflow](https://docs.github.com/en/get-started/quickstart/contributing-to-projects) is used for contributing. Below is a summary of the necessary steps for this workflow:
+
+1. Fork this repository.
+2. Clone the repository at your machine.
+3. Add your changes in a branch named after the feature (`lower-case-with-hyphens`).
+4. Make a pull request to `fRAT`, targeting the `master` branch.
+
 ## Images
 <img src="https://github.com/elliohow/fMRI_ROI_Analysis_Tool/blob/master/docs/images/GUI.png?raw=true" title="Example of the fRAT GUI" width=700>
 
