+++ docs/future_additions.rst
@@ -9,8 +9,6 @@ Future updates
 .. note::
    Any update ideas can be given here: https://github.com/elliohow/fMRI_ROI_Analysis_Tool/discussions
 
-   Any issues can be reported here: https://github.com/elliohow/fMRI_ROI_Analysis_Tool/issues
-
 General
 =======
 * Test fRAT with empty parameters for cases where a single metric is set to be reported and not a comparison between multiple parameters
