### 2023-03-06 12f5a96
* Bugfix: No longer errors when setting up table if there are no parameters to parse
