+++ fRAT/configuration_profiles/roi_analysis/fRAT_config.toml
@@ -1,10 +1,10 @@
 # Version Info
-version = '1.0.0'
+version = '1.0.1'
 
 # General
 run_analysis = true                                                               # true or false. Can skip this step if json files have already been created.
-run_statistics = true                                                             # true or false.
-run_plotting = true                                                               # true or false.
+run_statistics = false                                                            # true or false.
+run_plotting = false                                                              # true or false.
 verbose = true                                                                    # true or false.
 verbose_cmd_line_args = false                                                     # true or false.
 multicore_processing = true                                                       # true or false. Use multicore processing during analysis? Multicore processing currently works within participants not between them. Recommended: true
@@ -30,7 +30,7 @@ grey_matter_segment = true                                                      
 run_fsl_fast = 'Run if files not found'                                           # Recommended: "Run if files not found".  These files will only be searched for (and thus created) if "Use FSL FAST segmentation" is set to true. Options: ['Run if files not found', 'Never run'].
 fslfast_min_prob = 0.1                                                            # Recommended: 0.1
 stat_map_folder = ''                                                              # Folder name which contains the statistical map files. Example: temporalSNR_report
-stat_map_suffix = '_tSNR.nii.gz'                                                  # File name suffix of the statistical map files. Include the file extension. Example: _tSNR.img
+stat_map_suffix = '_tStd.nii.gz'                                                  # File name suffix of the statistical map files. Include the file extension. Example: _tSNR.img
 conf_level_number = '95%, 1.96'                                                   # Set the confidence level for confidence interval calculations. Numbers represent the confidence level and the corresponding critical z value. Recommended: 95%, 1.96. Options: ['80%, 1.28', '85%, 1.44', '90%, 1.64', '95%, 1.96', '98%, 2.33', '99%, 2.58'].
 binary_params = ['']                                                              # Add parameters here which will either be on or off.
 
@@ -48,12 +48,12 @@ minimum_voxels = 400                                                            
 bootstrap_samples = 1000                                                          # Recommended value 1000.  Note: Bootstrapping is only used to calculate percentage change versus baseline.
 bootstrap_confidence_interval = 95                                                # Recommended value: 95  Note: Bootstrapping is only used to calculate percentage change versus baseline.
 regional_stats_rois = ['all']                                                     # Provide a comma-separated list of regions, e.g. '3, 5', the string 'all' for all rois or the string 'Runtime' to provide regions at runtime.
-include_as_variable = ['INCLUDE ALL VARIABLES']                                   # Select which variables to include in statistical analysis. Used to determine which variables to use as fixed effects in linear mixed models and which variables to take into account when balancing data for the main effect t test data.
+include_as_variable = ['MB', 'SENSE']                                             # Select which variables to include in statistical analysis. Used to determine which variables to use as fixed effects in linear mixed models and which variables to take into account when balancing data for the main effect t test data.
 brain_map_p_thresh = 0.05                                                         # P-value threshold to use when creating brain coefficient maps. Any fixed effect that doesn't have a p-value equal to or less than this value will not be included in the coefficient map. 
 
 ## T-tests
 run_t_tests = true                                                                # true or false. 
-IV_type = ['FILL IV TYPE AS BETWEEN-SUBJECTS']                                    # Type of variable collected. Used to choose which t-test to use for pairwise comparisons.
+IV_type = ['Between-subjects', 'Between-subjects']                                # Type of variable collected. Used to choose which t-test to use for pairwise comparisons.
 
 ## Linear mixed models
 run_linear_mixed_models = true                                                    # true or false. 
@@ -66,8 +66,8 @@ interaction_effects = false                                                     
[…]
