+++ fRAT/configuration_profiles/maps/default_config.toml
@@ -1,5 +1,5 @@
 # Version Info
-version = '1.0.0'
+version = '1.0.1'
 
 # Statistical_maps
 
