+++ fRAT/configuration_profiles/maps/test_config.toml
@@ -1,5 +1,5 @@
 # Version Info
-version = '1.0.1'
+version = '1.1.0'
 
 # Statistical_maps
 
