### 2023-03-10 elliohow/fMRI_ROI_Analysis_Tool#35 @ZeitgeberH
[JOSS review] Missing data file from subject_example_data
Hi,
I was able to run fRAT under WSL2 in Windows 10 with functioning FSL.
As I go through the `Checking installation` step in the documentation,  it run into the `FileNotFoundError` error:
_FileNotFoundError: No such file or directory '/home/mlk/Documents/fRAT/subject_example_data/sub-02/statmaps/test_maps/P13_MB1_S2.5_match_BW_outliers.txt' for output 'out_file' of a MotionOutliers interface_

### Terminal output is as follows:

(frat) mlk@MingPC:~$ fRAT
fRAT 1.1.0 is developed and tested with Python 3.10.0.

INFO: Python version 3.10.9 is untested. Consider changing to version 3.10.0 if there are errors running fRAT.
----------------------------
----- Running fRAT_GUI -----
----------------------------
----- Running tests -----
----- Saved config file: fRAT_config.toml -----

fRAT 1.1.0 is developed and tested with Python 3.10.0.

INFO: Python version 3.10.9 is untested. Consider changing to version 3.10.0 if there are errors running fRAT.

--------------------------------
--- Statistical map creation ---
--------------------------------

Creating Temporal SNR maps.

Found 3 participant folders.

Starting processing pool using 24 cores.

Searching in folder: func
Saving output in directory: statmaps/test_maps

Creating statistical maps for participant: sub-02
      Creating statmaps for 16 files
        Analysing file: P14_MB2_S2.5_match_BW
        Analysing file: P1_MB3_S2_matchBW
        Analysing file: P7_MB2_S2_matchBW
        Analysing file: P13_MB1_S2.5_match_BW
        Analysing file: P3_MB3_S1.5
        Analysing file: P11_MB3_S1_match_BW
        Analysing file: P4_MB4_S1.5
        Analysing file: P2_MB1_S2_match_BW
        Analysing file: P16_MB4_S2.5_match_BW
        Analysing file: P9_MB1_S1_match_BW
        Analysing file: P15_MB3_S2.5_match_BW
        Analysing file: P5_MB4_S2_matchBW
        Analysing file: P8_MB1_S1.5_match_BW
        Analysing file: P6_MB2_S1.5
        Analysing file: P10_MB2_S1_match_BW
        Analysing file: P12_MB4_S1_match_BW
230310-11:54:12,852 nipype.interface INFO:
         stderr 2023-03-10T11:54:12.851991:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:12,954 nipype.interface INFO:
         stderr 2023-03-10T11:54:12.954617:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,163 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.162933:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,326 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.325954:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,368 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.367933:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,368 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.367949:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,371 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.371622:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,391 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.391104:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,468 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.468497:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,549 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.549016:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,550 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.550601:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,581 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.581737:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,593 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.593273:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,593 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.593283:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,593 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.593288:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230310-11:54:13,594 nipype.interface INFO:
         stderr 2023-03-10T11:54:13.594474:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
Exception in Tkinter callback
multiprocess.pool.RemoteTraceback:
"""
Traceback (most recent call last):
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/nipype/interfaces/base/core.py", line 454, in aggregate_outputs
    setattr(outputs, key, val)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/nipype/interfaces/base/traits_extension.py", line 330, in validate
    value = super(File, self).validate(objekt, name, value, return_pathlike=True)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/nipype/interfaces/base/traits_extension.py", line 135, in validate
    self.error(objekt, name, str(value))
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/traits/base_trait_handler.py", line 74, in error
    raise TraitError(
traits.trait_errors.TraitError: The 'out_file' trait of a MotionOutliersOutputSpec instance must be a pathlike object or string representing an existing file, but a value of '/home/mlk/Documents/fRAT/subject_example_data/sub-02/statmaps/test_maps/P13_MB1_S2.5_match_BW_outliers.txt' <class 'str'> was specified.

Traceback (most recent call last):
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/multiprocess/pool.py", line 125, in worker
    result = (True, func(*args, **kwds))
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/multiprocess/pool.py", line 51, in starmapstar
    return list(itertools.starmap(args[0], args[1]))
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/utils/statmap.py", line 248, in process_files_for_statmaps
    file, noise_file, redundant_files, outliers = prepare_statmap_files(file, no_ext_file, output_folder, participant)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/utils/statmap.py", line 190, in prepare_statmap_files
    file, outlier_timepoints, outlier_files = remove_motion_outliers(file, no_ext_file, output_folder, participant)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/utils/statmap.py", line 217, in remove_motion_outliers
    out_metric_values=outlier_values).run()
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/nipype/interfaces/base/core.py", line 401, in run
    outputs = self.aggregate_outputs(runtime)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/nipype/interfaces/base/core.py", line 461, in aggregate_outputs
    raise FileNotFoundError(msg)
FileNotFoundError: No such file or directory '/home/mlk/Documents/fRAT/subject_example_data/sub-02/statmaps/test_maps/P13_MB1_S2.5_match_BW_outliers.txt' for output 'out_file' of a MotionOutliers interface
"""

Traceback (most recent call last):
  File "/home/mlk/anaconda3/envs/frat/lib/python3.10/tkinter/__init__.py", line 1921, in __call__
    return self.func(*args)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/__main__.py", line 392, in <lambda>
    self.__setattr__(name, ttk.Button(self.settings_frame, command=lambda: func(self)))
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/__main__.py", line 845, in run_tests
    statmap_calc('Temporal SNR', __version__, 'test_config.toml', path_to_example_data)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/utils/statmap.py", line 396, in main
    calculate_statistical_maps(participants, output_folder, file_location, func)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/utils/statmap.py", line 281, in calculate_statistical_maps
    outliers = list(pool.starmap(process_files_for_statmaps, iterable))
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/multiprocess/pool.py", line 375, in starmap
    return self._map_async(func, iterable, starmapstar, chunksize).get()
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/multiprocess/pool.py", line 774, in get
    raise self._value
FileNotFoundError: No such file or directory '/home/mlk/Documents/fRAT/subject_example_data/sub-02/statmaps/test_maps/P13_MB1_S2.5_match_BW_outliers.txt' for output 'out_file' of a MotionOutliers interface
`
