### 2023-03-10 elliohow/fMRI_ROI_Analysis_Tool#35 @elliohow
Does the same error occur if you use python 3.10 to run fRAT?
