### 2023-03-10 elliohow/fMRI_ROI_Analysis_Tool#35 @elliohow
Have any files been saved in the test_maps folder for subject sub-02?
