### 2023-03-10 elliohow/fMRI_ROI_Analysis_Tool#35 @ZeitgeberH
Yeah. A few.  But no `motion_corrected` ,  `tMean`,  `tSNR`,` tStd` files

> ls ~/Documents/fRAT/subject_example_data/sub-02/statmaps/test_maps/ -ls
total 2388
164 -rw-r--r-- 1 mlk mlk 166547 Mar 10 12:58 P10_MB2_S1_match_BW_noise_volume.nii.gz
144 -rw-r--r-- 1 mlk mlk 143972 Mar 10 12:58 P11_MB3_S1_match_BW_noise_volume.nii.gz
144 -rw-r--r-- 1 mlk mlk 146581 Mar 10 12:58 P12_MB4_S1_match_BW_noise_volume.nii.gz
132 -rw-r--r-- 1 mlk mlk 133359 Mar 10 12:58 P13_MB1_S2.5_match_BW_noise_volume.nii.gz
172 -rw-r--r-- 1 mlk mlk 172303 Mar 10 12:58 P14_MB2_S2.5_match_BW_noise_volume.nii.gz
168 -rw-r--r-- 1 mlk mlk 169695 Mar 10 12:58 P15_MB3_S2.5_match_BW_noise_volume.nii.gz
180 -rw-r--r-- 1 mlk mlk 182529 Mar 10 12:58 P16_MB4_S2.5_match_BW_noise_volume.nii.gz
164 -rw-r--r-- 1 mlk mlk 165655 Mar 10 12:58 P1_MB3_S2_matchBW_noise_volume.nii.gz
128 -rw-r--r-- 1 mlk mlk 127140 Mar 10 12:58 P2_MB1_S2_match_BW_noise_volume.nii.gz
156 -rw-r--r-- 1 mlk mlk 157602 Mar 10 12:58 P3_MB3_S1.5_noise_volume.nii.gz
148 -rw-r--r-- 1 mlk mlk 148810 Mar 10 12:58 P4_MB4_S1.5_noise_volume.nii.gz
160 -rw-r--r-- 1 mlk mlk 162152 Mar 10 12:58 P5_MB4_S2_matchBW_noise_volume.nii.gz
128 -rw-r--r-- 1 mlk mlk 129834 Mar 10 12:58 P6_MB2_S1.5_noise_volume.nii.gz
168 -rw-r--r-- 1 mlk mlk 168002 Mar 10 12:58 P7_MB2_S2_matchBW_noise_volume.nii.gz
128 -rw-r--r-- 1 mlk mlk 128268 Mar 10 12:58 P8_MB1_S1.5_match_BW_noise_volume.nii.gz
 96 -rw-r--r-- 1 mlk mlk  94950 Mar 10 12:58 P9_MB1_S1_match_BW_noise_volume.nii.gz
  8 -rw-r--r-- 1 mlk mlk   4275 Mar 10 12:58 statmap_config.toml
