### 2023-03-16 03f50e9
* Bugfix: example data output folder now not static
