+++ fRAT/configuration_profiles/maps/default_config.toml
@@ -1,5 +1,5 @@
 # Version Info
-version = '1.1.1'
+version = '1.1.2'
 
 # Statistical_maps
 
