### 2023-03-16 elliohow/fMRI_ROI_Analysis_Tool#35 @elliohow
I've updated the example data files (https://osf.io/pbm3d/), the [installation instructions](https://fmri-roi-analysis-tool.readthedocs.io/en/latest/installation.html#on-linux-and-wsl2) for linux and WSL2, and made some changes that might fix this issue. Are you able to try the latest version following the new instructions to see if this fixes the issue?
