### 2023-03-16 elliohow/fMRI_ROI_Analysis_Tool#35 @ZeitgeberH
Hi @elliohow
I re-downloaded the data and updated fRAT to new version. It looks like the problem persist.
See below.
I'll try this on a Linux system this weekend, and report back what I found.

$ fRAT

fRAT 1.1.2 is developed and tested with Python 3.10.

----------------------------
----- Running fRAT_GUI -----
----------------------------
----- Running tests -----
----- Saved config file: fRAT_config.toml -----

fRAT 1.1.2 is developed and tested with Python 3.10.

--------------------------------
--- Statistical map creation ---
--------------------------------

Creating Temporal SNR maps.

Found 3 participant folders.

Starting processing pool using 24 cores.

Searching in folder: func
Saving output in directory: statmaps/test_maps

Creating statistical maps for participant: sub-02
      Creating statmaps for 16 files
        Analysing file: P14_MB2_S2.5_match_BW
        Analysing file: P1_MB3_S2_matchBW
        Analysing file: P7_MB2_S2_matchBW
        Analysing file: P13_MB1_S2.5_match_BW
        Analysing file: P3_MB3_S1.5
        Analysing file: P11_MB3_S1_match_BW
        Analysing file: P4_MB4_S1.5
        Analysing file: P2_MB1_S2_match_BW
        Analysing file: P16_MB4_S2.5_match_BW
        Analysing file: P9_MB1_S1_match_BW
        Analysing file: P15_MB3_S2.5_match_BW
        Analysing file: P5_MB4_S2_matchBW
        Analysing file: P8_MB1_S1.5_match_BW
        Analysing file: P6_MB2_S1.5
        Analysing file: P10_MB2_S1_match_BW
        Analysing file: P12_MB4_S1_match_BW
230316-14:15:03,487 nipype.interface INFO:
         stderr 2023-03-16T14:15:03.487089:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:03,690 nipype.interface INFO:
         stderr 2023-03-16T14:15:03.690732:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:03,772 nipype.interface INFO:
         stderr 2023-03-16T14:15:03.772377:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:03,851 nipype.interface INFO:
         stderr 2023-03-16T14:15:03.851643:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:03,851 nipype.interface INFO:
         stderr 2023-03-16T14:15:03.851672:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:03,851 nipype.interface INFO:
         stderr 2023-03-16T14:15:03.851677:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:03,987 nipype.interface INFO:
         stderr 2023-03-16T14:15:03.986986:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:04,6 nipype.interface INFO:
         stderr 2023-03-16T14:15:04.005910:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:04,7 nipype.interface INFO:
         stderr 2023-03-16T14:15:04.007771:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:04,101 nipype.interface INFO:
         stderr 2023-03-16T14:15:04.100986:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:04,114 nipype.interface INFO:
         stderr 2023-03-16T14:15:04.114405:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:04,176 nipype.interface INFO:
         stderr 2023-03-16T14:15:04.175970:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:04,176 nipype.interface INFO:
         stderr 2023-03-16T14:15:04.175970:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:04,176 nipype.interface INFO:
         stderr 2023-03-16T14:15:04.175996:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:04,177 nipype.interface INFO:
         stderr 2023-03-16T14:15:04.177502:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
230316-14:15:04,294 nipype.interface INFO:
         stderr 2023-03-16T14:15:04.294509:/home/mlk/fsl/bin/fsl_motion_outliers: 1: bc: not found
Exception in Tkinter callback
multiprocess.pool.RemoteTraceback:
"""
Traceback (most recent call last):
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/nipype/interfaces/base/core.py", line 454, in aggregate_outputs
    setattr(outputs, key, val)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/nipype/interfaces/base/traits_extension.py", line 330, in validate
    value = super(File, self).validate(objekt, name, value, return_pathlike=True)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/nipype/interfaces/base/traits_extension.py", line 135, in validate
    self.error(objekt, name, str(value))
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/traits/base_trait_handler.py", line 74, in error
    raise TraitError(
traits.trait_errors.TraitError: The 'out_file' trait of a MotionOutliersOutputSpec instance must be a pathlike object or string representing an existing file, but a value of '/home/mlk/Documents/fRAT/subject_example_data/sub-02/statmaps/test_maps/P13_MB1_S2.5_match_BW_outliers.txt' <class 'str'> was specified.

Traceback (most recent call last):
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/multiprocess/pool.py", line 125, in worker
    result = (True, func(*args, **kwds))
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/multiprocess/pool.py", line 51, in starmapstar
    return list(itertools.starmap(args[0], args[1]))
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/utils/statmap.py", line 253, in process_files_for_statmaps
    file, noise_file, redundant_files, outliers = prepare_statmap_files(file, no_ext_file, output_folder, participant)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/utils/statmap.py", line 190, in prepare_statmap_files
    file, outlier_timepoints, outlier_files = remove_motion_outliers(file, no_ext_file, output_folder, participant)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/utils/statmap.py", line 222, in remove_motion_outliers
    out_metric_values=outlier_values).run()
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/nipype/interfaces/base/core.py", line 401, in run
    outputs = self.aggregate_outputs(runtime)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/nipype/interfaces/base/core.py", line 461, in aggregate_outputs
    raise FileNotFoundError(msg)
FileNotFoundError: No such file or directory '/home/mlk/Documents/fRAT/subject_example_data/sub-02/statmaps/test_maps/P13_MB1_S2.5_match_BW_outliers.txt' for output 'out_file' of a MotionOutliers interface
"""

Traceback (most recent call last):
  File "/home/mlk/anaconda3/envs/frat310/lib/python3.10/tkinter/__init__.py", line 1921, in __call__
    return self.func(*args)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/__main__.py", line 384, in <lambda>
    self.__setattr__(name, ttk.Button(self.settings_frame, command=lambda: func(self)))
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/__main__.py", line 837, in run_tests
    statmap_calc('Temporal SNR', __version__, 'test_config.toml', path_to_example_data_input)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/utils/statmap.py", line 401, in main
    calculate_statistical_maps(participants, output_folder, file_location, func)
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/fRAT/utils/statmap.py", line 286, in calculate_statistical_maps
    outliers = list(pool.starmap(process_files_for_statmaps, iterable))
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/multiprocess/pool.py", line 375, in starmap
    return self._map_async(func, iterable, starmapstar, chunksize).get()
  File "/home/mlk/.local/pipx/venvs/frat-brain/lib/python3.10/site-packages/multiprocess/pool.py", line 774, in get
    raise self._value
FileNotFoundError: No such file or directory '/home/mlk/Documents/fRAT/subject_example_data/sub-02/statmaps/test_maps/P13_MB1_S2.5_match_BW_outliers.txt' for output 'out_file' of a MotionOutliers interface
