### 2023-03-16 elliohow/fMRI_ROI_Analysis_Tool#35 @elliohow
I can't replicate this issue. Although as it can create the files that don't require FSL, but it can't create any files that do, it might be an FSL setup issue.
