### 2023-03-16 #5200 @ZeitgeberH
Hi @elliohow
Could you add a section in your software paper to satisfy this requirement "State of the field: Do the authors describe how this software compares to other commonly-used packages?".
For example,  I think you could elaborate the last point towards the end of your abstract "There is currently a lack of easy-to-use tools to _analyse_ these metrics simultaneously across multiple ROI".
