### 2023-03-20 1994dd2
* Minor change: Made GUI more Linux compatible * Bugfix: Output data folder for installation testing now returns None if not available
