### 2023-03-20 acd1751
* Bugfix: Output data folder path now correct for installation tests
