### 2023-03-22 #5200 @billbrod
> Hi @elliohow Could you add a section in your software paper to satisfy this requirement "State of the field: Do the authors describe how this software compares to other commonly-used packages?". For example, I think you could elaborate the last point towards the end of your abstract "There is currently a lack of easy-to-use tools to _analyse_ these metrics simultaneously across multiple ROI".

Yes, I want to second this. In particular -- to what extent is fRAT an infrastructure / facilitation tool (like Nipype or fMRIPrep), making it easier to correctly run analyses that can be already be done using existing tools, and to what extent is it implementing novel (analysis, visualization, statistical) methods?
