### 2023-03-22 #5200 @billbrod
@mstimberg , this is more of a question about JOSS (since this is my first review) -- does JOSS allow / encourage explicit descriptions of author contributions? @elliohow has written (essentially) all the code and, following standard procedure in neuroscience/psychology, I'm thus assuming the other two authors are supervisors / involved in the conception of the package, but that's not made explicit anywhere. Should it be?
