### 2023-03-22 elliohow/fMRI_ROI_Analysis_Tool#36 @billbrod
[JOSS review] Documentation statement of need
- [ ] The JOSS paper includes a statement of need, but there is none included in the documentation. The version from the paper should be included (and possibly expanded upon).
- [ ] For both, I think a fuller description of what scientific approaches are enabled or facilitated by this package (with citations, if they already exist) would be good. Or what problems it helps avoid. Basically, it's not clear to me why a user would want to use ROI analysis over the standard voxelwise approach.

https://github.com/openjournals/joss-reviews/issues/5200
