### 2023-03-22 elliohow/fMRI_ROI_Analysis_Tool#37 @billbrod
[JOSS Review] Any citations using this method or demonstrating its need?
Are there any citations that could be included in the JOSS paper that use ROI-level analysis, discuss it, or demonstrate the need / utility of it (e.g., showing problems with voxelwise methods)?

https://github.com/openjournals/joss-reviews/issues/5200
