### 2023-03-22 elliohow/fMRI_ROI_Analysis_Tool#38 @billbrod
[JOSS Review] Tests?
Are there any tests beyond the [checking installation](https://fmri-roi-analysis-tool.readthedocs.io/en/latest/installation.html#checking-installation) sections of the documentation? Are there any automated tests?

I understand the package is focused on the GUI, which might be difficult to develop automated tests for, but would it be possible to put together tests to make sure e.g., installation completes and a sample analysis runs?

https://github.com/openjournals/joss-reviews/issues/5200
