### 2023-03-22 elliohow/fMRI_ROI_Analysis_Tool#38 @elliohow
Yes the checking installation section details how to run a sample analysis and if the "full comparison" option is checked, iit will also check that no files are missing between the sample analysis and the example dataset.
