### 2023-03-22 elliohow/fMRI_ROI_Analysis_Tool#39 @billbrod
[JOSS Review] Community guidelines
Instructions for opening an issue or getting help are present, but there's nothing for how to contribute. I would like to see a code of conduct and contributing guide. Here's a template from [the Turing Way](https://github.com/alan-turing-institute/reproducible-project-template), and some examples: [matplotlib](https://matplotlib.org/stable/devel/contributing.html#contributing), [nipype](https://github.com/nipy/nipype/blob/master/CONTRIBUTING.md)

https://github.com/openjournals/joss-reviews/issues/5200
