### 2023-03-22 elliohow/fMRI_ROI_Analysis_Tool#35 @billbrod
I was able to get the test to run using Linux (Fedora 36), FSL version 6.0, python 3.10, fRAT 1.1.3.
