### 2023-03-22 elliohow/fMRI_ROI_Analysis_Tool#38 @billbrod
Gotcha, so there are no other tests then? I would like to see some automated tests to check at least that the package installs and that a reduced version of the sample analysis (see below) runs. Would that be possible to add?

Additionally:
- [ ] When the test gets started, it checks whether there's anything other than `subject_example_data` exists in the `fRAT` directory. If there is, it raises an error (screenshot below, ignore the random white box). In this case, there were just two zip files in that directory (the ones downloaded from the OSF), not other subjects. Docs should probably make clear that nothing else is allowed in that directory while tests are being run.

![image](https://user-images.githubusercontent.com/6338509/227045906-3044e5c3-3445-422b-a504-f663cb00e120.png)

- [ ] What happens if "full comparison" is not checked? It just checks that it runs? (From reading the instructions, it would be easy to miss that "full comparison" should be checked, I almost did).
- [ ] Is the comparison just that the same files are present? Do you use something like `cmp` to test whether the files are identical? It would be good to check that they're identical (or, at least, good enough, if exact reproducibility cannot be guaranteed due to stochasticity or precision issues).
- [ ] That sample analysis takes a fairly large download (2.3GB for input, another 3.3GB for the check) and a while to run (>90 min, used all 12 of my laptop's cores for a large amount of it). Is it possible to include a more reduced version just to check that the install succeeded? For example, only a single subject, 1 or 2 ROIs, etc.
- [ ] It would also be nice to give users an expectation of how long / how much space / how many resources it will.
- [ ] It's unclear if the user should look at the output of the sample analysis (and if so, what they should look for) or whether the fact that it runs to completion is sufficient. Could you update the documentation to provide a bit of guidance there?
