### 2023-03-22 elliohow/fMRI_ROI_Analysis_Tool#38 @billbrod
Actually, my laptop shut down about 90 minutes into the sample analysis (I had walked away from my computer, so I don't know exactly where it was; not sure it was related to this). Do logs get created somewhere so I can check where it was? Is it possible to resume from where it was?
