### 2023-03-23 13d7993
* Minor change: Path to FSL not checked when first running GUI
