### 2023-03-23 elliohow/fMRI_ROI_Analysis_Tool#40 @elliohow
> * [ ]  Your [linux install](https://fmri-roi-analysis-tool.readthedocs.io/en/latest/installation.html#on-linux-and-wsl2) instructions tell people to modify their system python and use it to install fRAT. This [is a bad idea](https://askubuntu.com/questions/1398568/installing-python-who-is-deadsnakes-and-why-should-i-trust-them). It appears that users just need a working `pipx` install with python 3.10 available, so I believe it's better to tell them that explicitly, and point them to [miniconda](https://mamba.readthedocs.io/en/latest/installation.html) or [mamba](https://mamba.readthedocs.io/en/latest/installation.html) if they need a new install.

This method does not change the default system python, it just makes python 3.10 available. `python3` still calls the default system python, whereas to run python 3.10, it needs to called explicitly with: `python3.10`.

> * [ ]  It would be nice to check for FSL explicitly and raise a helpful exception when fRAT is started or at the beginning of a run (e.g., check for the environmental variable `$FSLDIR` and that `$FSLDIR/bin` is on user's `PATH`). Right now, the sample analysis starts running until the following error:
>       ![image](https://user-images.githubusercontent.com/6338509/227052756-bc24fcbe-4f1e-434f-8e30-a9b374931b82.png)

This has now been added in version 1.2.0.

> * [ ]  Your install instructions will only install this standalone version of the application in an isolated environment. Is that the only way you intend for users to make use of it? This would prevent users from, e.g., having a single environment with all their analysis packages in it. It's fine if so, just wanted to check.
>
> * [ ]  If you would like users to install fRAT along other packages (e.g., not just in an isolated environment with `pipx`), then I don't think pinning the specific versions of your dependencies is a good idea (it makes it hard to find e.g., a version of `pillow` that works with fRAT and other packages the user wants to install). But if it's just in an isolated environment, it probably doesn't matter.

I did consider only setting a minimum version for dependencies, but when upgrading dependencies I came across problems with deprecated functions and libraries that no longer worked together, therefore I chose to instead pin the dependencies and intend for users to only use fRAT in an isolated environment. venv or pyenv for the isolated environment was a consideration, however as fRAT is an application not a library, I thought pipx may be the best option to go with. Especially as it doesn't require the user to activate a virtual environment before running fRAT.
