### 2023-03-23 elliohow/fMRI_ROI_Analysis_Tool#40 @billbrod
> This method does not change the default system python. After installing python 3.10 in this way,`python3` still calls the default system python, whereas to run python 3.10, it needs to called explicitly with: `python3.10`.

My bad! This looks good to me then. I'd prefer to see something in the docs saying "you need 3.10 on your path and here is one way to do so" or the like (in my experience, many users find python environments and having multiple installs very confusing), but this is fine otherwise.

> This has now been added in version 1.2.0.

Great!

> I did consider only setting a minimum version for dependencies, but when upgrading dependencies I came across problems with deprecated functions and libraries that no longer worked together, therefore I chose to instead pin the dependencies and intend for users to only use fRAT in an isolated environment. venv or pyenv for the isolated environment was a consideration, however as fRAT is an application not a library, I thought pipx may be the best option to go with. Especially as it doesn't require the user to activate a virtual environment before running fRAT.

Okay, this sounds good to me then. I agree the `pipx` install is relatively straightforward.
