### 2023-03-23 elliohow/fMRI_ROI_Analysis_Tool#39 @billbrod
It's not necessary, but the discussion in #35 reminded me: it might be useful to define a bug report [issue template](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository) that will prompt users for a  lot of the basic information: what fRAT version, what FSL version, what python version, attach a screenshot or pasted version of the terminal output ,etc.
