### 2023-03-23 elliohow/fMRI_ROI_Analysis_Tool#39 @billbrod
I think that's a good start, but I think a code of conduct (you can lightly adapt [this one](https://www.contributor-covenant.org/)) is necessary.

And, while not necessary, the following would be nice to add to the contributing section: what kind of contributions you're interested and uninterested in (more documentation, examples, new features, integration with other tools, etc.), any info about coding style or API, and required documentation and/or tests. Basically, what should they include in the PR? E.g., a description of what it includes, a small working example, the need it fills, etc.
