### 2023-03-23 elliohow/fMRI_ROI_Analysis_Tool#38 @elliohow
> Actually, my laptop shut down about 90 minutes into the sample analysis (I had walked away from my computer, so I don't know exactly where it was; not sure it was related to this). Do logs get created somewhere so I can check where it was? Is it possible to resume from where it was?

As shown in the flowchart in the paper, there are 3 main steps to the analysis: creation of statistical maps, ROI analysis, and figure creation & statistics. At the start of each step, a file (such as "analysis_log.toml") is created showing the settings used for that step, however there is not a log showing where the program ran up to. Based on what folders have been created, it is possible to see which steps have been created though. Each step can be ran independently in the normal analysis given the correct inputs, however this is not possible in the sample analysis without manually updating the test_config.toml file. The terminal output for the statistics step is already saved to easily find the results for a speciic ROI, but I will add onto the future additions page of the documentation to save a log of all terminal output for cases such as this.
