### 2023-03-23 elliohow/fMRI_ROI_Analysis_Tool#38 @elliohow
> * [ ]  Is the comparison just that the same files are present? Do you use something like `cmp` to test whether the files are identical? It would be good to check that they're identical (or, at least, good enough, if exact reproducibility cannot be guaranteed due to stochasticity or precision issues).

Originally I was using `filecmp` to test whether files are identical, however the byte-by-byte comparison of `filecmp` flagged many files as different due to precision issues. So instead I changed to just checking that all files in the example data were also created by the sample analysis.
