### 2023-03-23 elliohow/fMRI_ROI_Analysis_Tool#38 @elliohow
* [ ]  What happens if "full comparison" is not checked? It just checks that it runs? (From reading the instructions, it would be easy to miss that "full comparison" should be checked, I almost did).
>
>  * [ ]  That sample analysis takes a fairly large download (2.3GB for input, another 3.3GB for the check) and a while to run (>90 min, used all 12 of my laptop's cores for a large amount of it). Is it possible to include a more reduced version just to check that the install succeeded? For example, only a single subject, 1 or 2 ROIs, etc.
>
>  * [ ]  It would also be nice to give users an expectation of how long / how much space / how many resources it will.
>
>  * [ ]  It's unclear if the user should look at the output of the sample analysis (and if so, what they should look for) or whether the fact that it runs to completion is sufficient. Could you update the documentation to provide a bit of guidance there?

There are two folders on osf as one was just to test that the sample analysis ran to completion, whereas the other also needed to be downloaded to check whether any files were missing. The number of participants and files were chosen to stop issues with the linear mixed model flagging up convergence warnings due to the random effects covariance being singular, with the tradeoff being much larger file sizes. On balance, I agree that it would be less confusing to just have one folder to download, reduce the number of participants & files, and have the file comparison always be ran. Ill just update the documentation to warn that there will be warnings flagged at the statistics step.

I'm just creating a new 1.3 GB example_data folder to replace the two current ones. As there are only 4 files for for each participant, the analysis step will now only use a max of 4 cores (in the real analysis, the user is able to choose the number of cores to use). While it is not possible to reduced the number of ROIs during the analysis step as they are simultaneously analysed, i've reduced the number of ROIs to plot and calculate statistics for in the sample analysis to from all 50 to 2.
