### 2023-03-24 elliohow/fMRI_ROI_Analysis_Tool#35 @elliohow
I was able to run the test using Ubuntu 18.04 both natively and with WSL (FSL version 6.0, python 3.10, fRAT 1.2.0).
