### 2023-03-24 dc14071
* Documentation update: future additions page edited * Documentation update: installation page edited to streamline instructions for checking installation * Minor change: file comparison when checking installation is now automatic * Minor change: expected example data folder structure now different * Minor change: changed where some statistics files are saved
