+++ fRAT/__main__.py
@@ -147,12 +147,14 @@ class GUI:
 
         self.statmap_reset_button = ttk.Button(self.statmap_settings_frame)
         self.statmap_reset_button.place(relx=0.5, rely=0.08, height=42, width=105)
-        self.statmap_reset_button.configure(command=lambda: Reset_settings(['Statistical_maps']), text='''Reset settings''')
+        self.statmap_reset_button.configure(command=lambda: Reset_settings(['Statistical_maps']),
+                                            text='''Reset settings''')
         Tooltip.CreateToolTip(self.statmap_reset_button, 'Reset statistical map settings')
 
         self.statmap_settings_button = ttk.Button(self.statmap_settings_frame)
         self.statmap_settings_button.place(relx=0.17, rely=0.55, height=42, width=150)
-        self.statmap_settings_button.configure(command=lambda: self.change_frame('Statistical_maps'), text='''Settings''')
+        self.statmap_settings_button.configure(command=lambda: self.change_frame('Statistical_maps'),
+                                               text='''Settings''')
 
     def statmap_run_frame_create(self, window):
         self.statmap_run_frame = tk.LabelFrame(window)
@@ -197,7 +199,8 @@ class GUI:
                     try:
                         eval(curr_page)[setting[0]]['Current'] = current_setting
                     except KeyError:
-                        warnings.warn(f'"{setting[0]}" not present in config setup file. Configuration file may be outdated.')
+                        warnings.warn(
+                            f'"{setting[0]}" not present in config setup file. Configuration file may be outdated.')
 
     def banner_draw(self, window):
         img = Image.open(f'{Path(os.path.abspath(__file__)).parents[0]}/images/fRAT.gif')
@@ -676,7 +679,8 @@ class GUI:
                     else:
                         current_options2_menu[current_widget['label']] = []
 
-                    current_widget['Current'][len(current_options2_menu[current_widget['label']])] = self.dynamic_widgets[widget].val.get()
+                    current_widget['Current'][len(current_options2_menu[current_widget['label']])] = \
+                    self.dynamic_widgets[widget].val.get()
 
         for frame in self.frames:
             frame.destroy()
@@ -847,79 +851,59 @@ def run_tests(GUI):
     GUI.change_frame('General')
     Save_settings(pages, f'roi_analysis/{ConfigurationFiles.analysis_config}')
 
-    path_to_example_data_input, path_to_example_data_output = find_example_dataset()
+    path_to_example_data = find_example_dataset()
 
     # Create tSNR maps and run ROI analysis
-    statmap_calc('Temporal SNR', __version__, 'test_config.toml', path_to_example_data_input)
-    fRAT('test_config.toml', path_to_example_data_input)
+    statmap_calc('Temporal SNR', __version__, 'test_config.toml', path_to_example_data)
+    fRAT('test_config.toml', path_to_example_data)
 
     # Run tests to check if output of fRAT matches the example data
-    if General['full_comparison']['Current'] == 'true':
-        roi_output_test = TestDifferences([f'{path_to_example_data_input}/sub-02/statmaps/test_maps',
-                                            f'{path_to_example_data_input}/sub-02/statmaps/temporalSNR_report'],
-                                           General['verbose_errors']['Current'])
+    roi_output_test = TestDifferences([f'{path_to_example_data}/sub-02/statmaps/test_maps',
+                                       f'{path_to_example_data}/sub-02/statmaps/temporalSNR_report'],
+                                      General['verbose_errors']['Current'])
 
-        voxelwise_map_test = TestDifferences([f'{path_to_example_data_input}/test_ROI_report',
-                                              f'{path_to_example_data_output}'],
[…]
