+++ fRAT/configuration_profiles/maps/statmap_config.toml
@@ -1,5 +1,5 @@
 # Version Info
-version = '1.2.0'
+version = '1.3.0'
 
 # Statistical_maps
 
