+++ fRAT/configuration_profiles/roi_analysis/default_config.toml
@@ -1,5 +1,5 @@
 # Version Info
-version = '1.2.0'
+version = '1.3.0'
 
 # General
 run_analysis = true                                                               # true or false. Can skip this step if json files have already been created.

[…]
