### 2023-03-24 elliohow/fMRI_ROI_Analysis_Tool#38 @elliohow
Version 1.3.0 has now been released that should address these issues. The sample analysis now takes roughly 15 minutes, and the download is significantly smaller. File comparison is automatically ran, and so there is only one file to download from osf. Instructions given to run the sample analysis have also now been simplified and should be clearer.
