### 2023-03-24 d261e18
* Documentation update: installation page edited to include installation of bc for linux and WSL2 systems
