### 2023-03-24 elliohow/fMRI_ROI_Analysis_Tool#35 @elliohow
Thank you for the solution. I've added it to the installation page.
