### 2023-03-24 elliohow/fMRI_ROI_Analysis_Tool#39 @billbrod
Some guidance on [codes of conduct](https://mozilla.github.io/open-leadership-training-series/articles/building-communities-of-contributors/write-a-code-of-conduct/) and [contributor guidelines](https://mozilla.github.io/open-leadership-training-series/articles/building-communities-of-contributors/write-contributor-guidelines/) from Mozilla's open leadership program.
