### 2023-03-24 6d88dc0
* Minor change: Increased number of regional stats calculated during LMM to increase number of data points passed to compute_rsquare_regression function. This should prevent occurrence of an error due to only one ROI being passed to this function, which prevents a constant column being added to the data, and makes the implied shape of the data not match the passed shape.
