### 2023-03-24 75cd74c
* Minor change: Added check to make sure compute_rsquare_function can not receive any input of length 1
