+++ fRAT/configuration_profiles/maps/statmap_config.toml
@@ -1,5 +1,5 @@
 # Version Info
-version = '1.3.1'
+version = '1.3.2'
 
 # Statistical_maps
 
