+++ fRAT/configuration_profiles/roi_analysis/test_config.toml
@@ -1,10 +1,10 @@
 # Version Info
-version = '1.3.1'
+version = '1.3.2'
 
 # General
-run_analysis = true                                                               # true or false. Can skip this step if json files have already been created.
+run_analysis = false                                                               # true or false. Can skip this step if json files have already been created.
 run_statistics = true                                                             # true or false.
-run_plotting = true                                                               # true or false.
+run_plotting = false                                                               # true or false.
 verbose = true                                                                    # true or false.
 verbose_cmd_line_args = false                                                     # true or false.
 multicore_processing = true                                                       # true or false. Use multicore processing during analysis? Multicore processing currently works within participants not between them. Recommended: true
