### 2023-03-24 b3f9575
* Bugfix: reset test options
