### 2023-03-30 33130bd
* Bugfix: When saving log of settings, all subsections are now correctly saved * Bugfix: Resolved error with non-capitalised file path which caused errors for case-sensitive OSs * Bugfix: Solved broken pipe error for Linux using SIGPIPE * Minor change: version number removed from settings files as version information is automatically added when saving logs
