+++ fRAT/configuration_profiles/maps/default_config.toml
@@ -1,6 +1,3 @@
-# Version Info
-version = '1.3.3'
-
 # Statistical_maps
 
 ## General settings
