### 2023-03-30 #5200 @elliohow
@mstimberg Have had a bit of friction getting `fRAT` set up with Linux and WSL2 due to some unexpected bugs, but hopefully all fixed now (🤞).  I hope to have the remaining issues addressed by the end of next week.
