### 2023-03-30 #5200 @mstimberg
> @mstimberg Have had a bit of friction getting `fRAT` set up with Linux and WSL2 due to some unexpected bugs, but hopefully all fixed now (crossed_fingers). I hope to have the remaining issues addressed by the end of next week.

Great, thanks for the feedback :+1: .
