### 2023-04-03 4f34f52
* Minor change: Updated GUI text style to increase cross-platform compatibility (and hopefully solve a bug with bold not showing on some Linux distributions) * Minor change: Updated documentation and tooltips to make it clearer that any critical parameters are supported * Minor change: Increased clarity of what the "Atlas information" section on the GUI is for in the tutorial
