+++ fRAT/__main__.py
@@ -9,6 +9,7 @@ import textwrap
 
 from PIL import ImageTk
 from operator import itemgetter
+from tkinter import font
 
 from fRAT.nogui import fRAT
 from fRAT._version import __version__
@@ -28,7 +29,6 @@ signal(SIGPIPE, SIG_DFL)
 w = None
 WIDGET_Y_PADDING = 9
 WIDGET_X_PADDING = 10
-
 
 def start_gui():
     '''Starting point when module is the main routine.'''
@@ -76,6 +76,13 @@ class GUI:
             window.configure(highlightbackground=self.background)
             window.configure(highlightcolor="black")
 
+            self.default_font = None
+            self.important_font = None
+            self.subheading_font = None
+            self.heading_font = None
+
+            self.setup_fonts()
+
         self.change_page_specific_settings(window)
 
         if page == 'Statistical_maps':
@@ -95,26 +102,27 @@ class GUI:
             self.create_homestar_runner_frame(self.settings_frame)
             self.create_statmap_frame(window)
 
+    def setup_fonts(self):
+        self.default_font = font.Font(font='TkDefaultFont')
+        self.important_font = font.Font(font='TkDefaultFont')
+        self.heading_font = font.Font(font='TkDefaultFont')
+        self.subheading_font = font.Font(font='TkDefaultFont')
+
+        self.important_font.configure(weight='bold')
+        self.heading_font.configure(weight='bold', size=20)
+        self.subheading_font.configure(weight='bold', size=18)
+
     def change_page_specific_settings(self, window):
-        if platform.system() == 'Darwin':
-            statistics_width = "530"
-            home_width = "512"
-            statmap_width = "500"
-            general_width = "480"
-            parsing_width = "450"
-        else:
-            statistics_width = "550"
-            home_width = "530"
-            statmap_width = "520"
-            general_width = "500"
-            parsing_width = "480"
+        statistics_width = "530"
+        home_width = "512"
+        parsing_width = "450"
 
         if self.page == 'Home':
             window.geometry(f"{home_width}x860")
         elif self.page == 'General':
-            window.geometry(f"{general_width}x860")
+            window.geometry(f"{home_width}x860")
         elif self.page == 'Analysis':
-            window.geometry(f"{general_width}x860")
+            window.geometry(f"{home_width}x860")
         elif self.page == 'Parsing':
             window.geometry(f"{parsing_width}x280")
         elif self.page == 'Plotting':
@@ -122,14 +130,14 @@ class GUI:
         elif self.page == 'Statistics':
             window.geometry(f"{statistics_width}x860")
         elif self.page == 'Statistical_maps':
-            window.geometry(f"{statmap_width}x860")
+            window.geometry(f"{home_width}x860")
 
     def create_statmap_frame(self, window):
         self.statmap_frame = tk.LabelFrame(window)
         self.frames.append(self.statmap_frame)
 
         self.statmap_frame.place(relx=0.02, rely=0.77, relheight=0.22, relwidth=0.973)
-        self.statmap_frame.configure(text='Statistical maps', font='Helvetica 18 bold')
+        self.statmap_frame.configure(text='Statistical maps', font=self.heading_font)
         self.format_frame(self.statmap_frame, borderwidth=1)
 
         self.statmap_run_frame_create(self.statmap_frame)
@@ -139,7 +147,7 @@ class GUI:
         self.statmap_settings_frame = tk.LabelFrame(window)
         self.frames.append(self.statmap_settings_frame)
         self.statmap_settings_frame.place(relx=0.02, rely=0, height=140, relwidth=0.467)
[…]
