+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,31 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: "[BUG]"
+labels: bug
+assignees: ''
+
+---
+
+### Summary
+
+### Actual behavior
+
+### Expected behavior
+
+### How to replicate the behavior
+
+### Workflow details (if applicable)
+
+### Platform details:
+ - OS: [e.g. MacOS Catalina]
+ - Python version: [e.g. 3.10.2]
+ - fRAT version: [e.g. 1.3.4]
+ - FSL version: [e.g. 6.0.6]
+ - Container (if applicable): [e.g. Conda]
+
+### Terminal output (if applicable)
+ 
+### Screenshots (if applicable)
+
+### Additional context (if applicable)
