+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -7,14 +7,14 @@ assignees: ''
 
 ---
 
-**Is your feature request related to a problem? Please describe.**
-A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+### Is your feature request related to a problem? (If applicable)
+A clear and concise description of what the problem is. E.g. I'm always frustrated when [...]
 
-**Describe the solution you'd like**
+### Describe the solution you'd like
 A clear and concise description of what you want to happen.
 
-**Describe alternatives you've considered**
+### Describe alternatives you've considered (if applicable)
 A clear and concise description of any alternative solutions or features you've considered.
 
-**Additional context**
+### Additional context (if applicable)
 Add any other context or screenshots about the feature request here.
