+++ docs/index.rst
@@ -50,19 +50,13 @@ To learn how to run an ROI analysis, follow this :doc:`tutorial </tutorials/Basi
 
 Reporting bugs
 --------------
-To report a bug, please go to `fRAT's Issues <https://github.com/elliohow/fMRI_ROI_Analysis_Tool/issues/new>`_.
+To report a bug or suggest a new feature, please go to `fRAT's Issues <https://github.com/elliohow/fMRI_ROI_Analysis_Tool/issues/new>`_.
 
 For other questions, issues or discussion please go to `fRAT's Discussions <https://github.com/elliohow/fMRI_ROI_Analysis_Tool/discussions>`_.
 
 Contributing with development
 -----------------------------
-
-The `Fork & Pull Request Workflow <https://docs.github.com/en/get-started/quickstart/contributing-to-projects>`_ is used for contributing. Below is a summary of the necessary steps for this workflow:
-
-1. Fork this repository.
-2. Clone the repository at your machine.
-3. Add your changes in a branch named after the feature (`lower-case-with-hyphens`).
-4. Make a pull request to `fRAT`, targeting the `master` branch.
+If you'd like to contribute to the project please read our `contributing guidelines <https://github.com/elliohow/fMRI_ROI_Analysis_Tool/blob/master/CONTRIBUTING.md>`_. Please also read through our `code of conduct <https://github.com/elliohow/fMRI_ROI_Analysis_Tool/blob/master/CODE_OF_CONDUCT.md>`_.
 
 
 GUI images
