### 2023-04-04 elliohow/fMRI_ROI_Analysis_Tool#39 @elliohow
I've added a code of conduct, contributing guidelines and issue templates, and linked to them in the documentation.
