### 2023-04-05 29afd95
* Minor change: Made font size specific to platform * Documentation update: Gave information to fix conda text rendering issue
