+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -16,11 +16,11 @@ assignees: ''
 ### How to replicate the behavior
 
 ### Platform details:
- - OS: [e.g. MacOS Catalina]
+ - OS: [e.g. MacOS Catalina, Ubuntu 18.04 on WSL2, etc.]
  - Python version: [e.g. 3.10.2]
  - fRAT version: [e.g. 1.3.4]
  - FSL version: [e.g. 6.0.6]
- - Container (if applicable): [e.g. Conda]
+ - Package managers used (other than pip and pipx): [e.g. Conda]
 
 ### Terminal output (if applicable)
  
