### 2023-04-20 f97d657
* Bugfix: First attempt at forcing PIL to properly close images after use. * Bugfix: Int type forced if multiprocessing core count now manually chosen. * Paper change: Updated image
