### 2023-04-20 #5200 @ZeitgeberH
Hi @mstimberg and @elliohow
I have worked through the test and tutorial part.  The package  (version `1.3.6`) is now fully functional in my WSL2 environment .  From my side,  remaining issue is the improvement on the paper (`State of the field` and associated references).
