### 2023-04-24 elliohow/fMRI_ROI_Analysis_Tool#39 @billbrod
This looks good!
