### 2023-04-24 elliohow/fMRI_ROI_Analysis_Tool#38 @billbrod
The installation tests work for me without a problem! It might be worth adding a note to the section about the FSL requirement to say: "Make sure that `$FSLDIR` is set and `$FSLDIR/bin` is on your path" (and `$FSLOUTPUTTYPE` is set as well?). I didn't add that to my `.bashrc` and so had to do that manually. I don't think it's *necessary* because fRAT throws informative errors if it's an issue, but it might be nice to have it listed in the docs.

I'm going to close this issue now, and open another one to discuss automated tests.
