### 2023-04-24 elliohow/fMRI_ROI_Analysis_Tool#45 @billbrod
[JOSS Review] Automated tests?
I mentioned this briefly in #38 , but then we manually disccussed the installation tests, so I wanted to open a separate issue for this.

Have you thought about automated tests? I understand that fRAT is a GUI and so automated testing is more difficult, but some automated tests to make sure it installs and the installation tests run would be nice. What do you think?

https://github.com/openjournals/joss-reviews/issues/5200
