### 2023-04-24 elliohow/fMRI_ROI_Analysis_Tool#45 @elliohow
Isn't the test described on the [installation page](https://fmri-roi-analysis-tool.readthedocs.io/en/latest/installation.html#checking-installation-with-a-test-analysis) already testing whether installation has been successful? If there have been any issues with the  install, the test won't complete and there will be missing files.
