### 2023-04-24 elliohow/fMRI_ROI_Analysis_Tool#46 @billbrod
[JOSS Review] Functionality documentation and non-GUI use
The documentation currently does not include documentation of the code for fRAT, focusing on a (very thorough!) tutorial for how to use it. My guess is you plan to only encourage / support use through the GUI, is that correct? In that case, I think documentation of the underlying functions is not necessary, but I do think you should provide more documentation of the different options in the GUI (see the [BET user guide](https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/BET/UserGuide) for an example from FSL). Some of this information is included in the [tutorial](https://fmri-roi-analysis-tool.readthedocs.io/en/latest/tutorials/Basic-ROI-analysis.html), but it should be pulled out into its own page. For example, under `Parsing`, what are the possible critical parameter values (and actually, why is this a free text field, rather than a checklist? if I'm understanding it correctly). What is `SENSE` and is it an alternative or complementary to `Multiband`? Or, under `Analysis -> Outlier detection`, what does checking `Noise cutoff` do? Or `Gaussian outlier (GauO) detection`? It doesn't need to be exhaustive (for example, the various file name and most of the plotting fields are pretty self-explanatory), but the critical ones should be documented (and that can often be a one-line explanation with a reference or link for existing methods).

https://github.com/openjournals/joss-reviews/issues/5200
