### 2023-04-24 elliohow/fMRI_ROI_Analysis_Tool#45 @billbrod
Yes! But it doesn't look to me like this test is automated -- it will require you to manually run it to make sure any changes don't break it. I believe you can create an automated version of that test using [Github actions](https://docs.github.com/en/actions) (or another continuous integration service) which will run every time you push or before you merge a PR (which is especially helpful for testing changes from outside contributors).

The test on the installation page does everything you need, but it's the automation I want to highlight here -- does that make sense?
