### 2023-04-24 elliohow/fMRI_ROI_Analysis_Tool#46 @elliohow
> My guess is you plan to only encourage / support use through the GUI, is that correct?

Yes currently I only support use of fRAT through the GUI currently.

> but I do think you should provide more documentation of the different options in the GUI (see the [BET user guide](https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/BET/UserGuide) for an example from FSL). Some of this information is included in the [tutorial](https://fmri-roi-analysis-tool.readthedocs.io/en/latest/tutorials/Basic-ROI-analysis.html), but it should be pulled out into its own page. For example, under `Parsing`, what are the possible critical parameter values (and actually, why is this a free text field, rather than a checklist? if I'm understanding it correctly). What is `SENSE` and is it an alternative or complementary to `Multiband`? Or, under `Analysis -> Outlier detection`, what does checking `Noise cutoff` do? Or `Gaussian outlier (GauO) detection`? It doesn't need to be exhaustive (for example, the various file name and most of the plotting fields are pretty self-explanatory), but the critical ones should be documented (and that can often be a one-line explanation with a reference or link for existing methods).
>

There are tooltips that explain the function of the parameter settings when you hover over them in the GUI, but I definitely plan to expand on these in the future in the documentation. Is it sufficient for now to add this to the future additions page on the documentation and expand on what is currently there in the future?
