### 2023-04-24 elliohow/fMRI_ROI_Analysis_Tool#46 @billbrod
> Yes currently I only support use of fRAT through the GUI currently.

Sounds good.

> There are tooltips that explain the function of the parameter settings when you hover over them in the GUI, but I definitely plan to expand on these in the future in the documentation. Is it sufficient for now to add this to the future additions page on the documentation and expand on what is currently there in the future?

Oh I missed the hovering. In that case, yes, with two small steps:
- put something in your docs about how hovering over the name of the option gives additional info (not the entry field, which is why I missed it).
- put some information about `SENSE` and `Multiband` into the docs, with a link or reference, since they are important. Multiband I've heard of, but it's not clear to me what it means here (I tend to think of it in the context of acquisition, not analysis, but that might be me), and I've never heard of SENSE.
