### 2023-04-24 elliohow/fMRI_ROI_Analysis_Tool#46 @elliohow
I've expanded on what the critical parameters setting is in the ROI tutorial:

> Next, enter the Parsing page found under the “Settings” panel in the GUI. The critical parameters setting indicates what the independent variables of the experiment are and will be used when labelling plots and creating the parameter csv file, whereas the critical parameters abbreviation setting indicates how these variables are represented in the filenames of the scans (if at all). For example, the critical parameters option may contain “Multiband, SENSE”, whereas the “critical parameters abbreviation” option will instead contain “mb, s”. In this case, the file name P1_MB3_S2_matchBW.nii will represent that Multiband 3 and SENSE 2 were used.
>
> Note
>
> The critical parameter settings are used to supply the names and file name abbreviations of the independent variables, therefore fRAT supports the use of any parameters (and any number of them).
>
> If critical parameters is left blank, the ROI analysis will combine all datasets, regardless of the independent variable used for each file. This can be useful for calculating tSNR across the entire dataset, if the independent variable does not affect tSNR.
>
