### 2023-04-24 bce65d7
* Bugfix: Moved to using a non-GUI backend for matplotlib to prevent known memory leak issue caused by creating subplots in a loop
