+++ fRAT/configuration_profiles/maps/statmap_config.toml
@@ -13,8 +13,8 @@ temporal_filter = true                                                          
 highpass_filter_cutoff = 0.01                                                     # Highpass filter cutoff frequency converted into sigma in seconds using the formula 1/(2*f*TR). Recommended: 0.01
 
 ## Motion correction
-remove_motion_outliers = false                                                    # true or false. Use fsl_motion_outliers to remove motion outliers(uses default fsl_motion_outliers settings). Recommended: true
-motion_correction = false                                                         # true or false. Use MCFLIRT to motion correct volumes (uses default MCFLIRT settings). Recommended: true
+remove_motion_outliers = true                                                     # true or false. Use fsl_motion_outliers to remove motion outliers(uses default fsl_motion_outliers settings). Recommended: true
+motion_correction = true                                                          # true or false. Use MCFLIRT to motion correct volumes (uses default MCFLIRT settings). Recommended: true
 
 ## Spatial smoothing
 spatial_smoothing = false                                                         # true or false. Uses SUSAN to spatial smooth. Recommended: true
@@ -28,5 +28,5 @@ noise_volume_location = 'End'                                                   
 iSNR_std_use_only_nonzero_voxels = true                                           # true or false.
 
 ## Add Gaussian noise
-noise_multipliers = [1, 2, 5]                                                     # Provide a comma-separated list of multipliers for the standard deviation of the gaussian noise to plot e.g. '1, 5'. A separate file will be produced for each multiplier. NOTE: Noise has a gaussian distribution, with a mean of 0 and a standard deviation of the noise level of each participant * multiplier.
+noise_multipliers = 1                                                             # Provide a comma-separated list of multipliers for the standard deviation of the gaussian noise to plot e.g. '1, 5'. A separate file will be produced for each multiplier. NOTE: Noise has a gaussian distribution, with a mean of 0 and a standard deviation of the noise level of each participant * multiplier.
 
