### 2023-04-24 elliohow/fMRI_ROI_Analysis_Tool#38 @elliohow
What did you need to add to your bash profile? I've tested fRAT with both Ubuntu 18.04 and 22.04 (using WSL2) and haven't needed to add anything to my path. The fslinstaller set up those for me.
