### 2023-04-24 elliohow/fMRI_ROI_Analysis_Tool#45 @elliohow
Ah yes I understand what you mean now. To run automate the running of the tests, the runner would need to be able to setup FSL and be able to install the example data from OSF. I'm not too familiar with the capabilities of Github actions, is this possible? Theres also the [storage limits](https://docs.github.com/en/billing/managing-billing-for-github-actions/about-billing-for-github-actions) for a runner. Not sure if im misunderstanding but I wouldn't be able to set up github actions for my project without getting GitHub Pro since the example dataset is 1.1GB.
