### 2023-04-24 elliohow/fMRI_ROI_Analysis_Tool#45 @billbrod
Oh right, I forgot about FSL. I'm not sure there's a way to install FSL on a github runner that follows their license. You could see if / how nipype does it, I guess. And you're right about the storage limit as well: unless you can come up with an even smaller example dataset, you won't be able to fit it on the runner. You could also check other Continuous Integration services to see if any give you enough storage; it looks like [Circle CI](https://circleci.com/docs/using-linuxvm/) might give you enough, but I'm not familiar with them.

So all told, this looks like more work than I expected, and so I don't think this is required for JOSS. I'd encourage you to look into it a bit more, especially if you start getting PRs from outside contributors, because (speaking from experience) it does make maintenance much easier. So I'll go ahead and close this.
