### 2023-04-24 elliohow/fMRI_ROI_Analysis_Tool#46 @billbrod
Yes that's much clearer. I'd still pull some of this info out to a different part of the docs as well, so it's easy to refer to. Or at least make it clear that everyone should go through the tutorial. My default assumption would be that I don't need to look through it if I understand how ROI analysis works, but the tutorial also includes key information about the tool works. That is, I thought the tutorial was only introducing the scientific concepts, but it is also introducing technical ones. So changing the wording on the index to say "To learn how to use fRAT, see this tutorial", instead of "To learn
