+++ paper/paper.md
@@ -22,22 +22,25 @@ affiliations:
    index: 1
  - name: Sir Peter Mansfield Imaging Centre, University of Nottingham, Nottingham, United Kingdom
    index: 2
-date: 24 February 2023
+date: 26 April 2023
 bibliography: paper.bib
 ---
 
 # Summary
 
-Functional magnetic resonance imaging (fMRI) is widely used to address basic cognitive as well as clinical neuroscience questions. The specific choice of imaging sequence and parameters used for image acquisition, can have a marked effect on the acquired data. Because the effects of different parameter settings, such as echo time (TE) or parallel imaging acceleration factors, can vary spatially and may also interact [@clareSingleshotMeasurementEstablish2001], assessing how these metrics affect data quality over brain regions is crucial. Such data quality metrics, in particular the temporal signal-to-noise ratio (tSNR), can therefore be used to optimise MRI scan parameters for a chosen region-of-interest (ROI). Reporting of these metrics over either the whole brain or for a small number of voxels is common; however this can obscure important inter-regional differences. There is currently a lack of easy-to-use tools to analyse these metrics simultaneously across multiple ROIs. The goal of the fMRI ROI Analysis Tool (`fRAT`), presented here, is to provide a toolset to address this gap, and provide a straightforward method for conducting ROI-wise
-analyses of fMRI metrics.
+Functional magnetic resonance imaging (fMRI) is widely used to address basic cognitive as well as clinical neuroscience questions. The specific choice of imaging sequence and parameters used for image acquisition can have a marked effect on the acquired data. The choice of different parameter settings used for the echo-planar imaging (EPI) acquisition include the trade-offs between an increase in-plane parallel acceleration factor (SENSE, GRAPPA, ASSET) to reduce geometric distortions and increase in Multiband (MB) factor to reduce scan time, both of which lead to spatially dependent noise amplification. In addition, the T2<sup>*</sup> varies across the brain influencing the optimal echo time (TE)[@clareSingleshotMeasurementEstablish2001]. Assessing how these parameters affect data quality over brain regions is crucial. Such data quality metrics, in particular temporal signal-to-noise ratio (tSNR) of the fMRI scan, can be used to optimise fMRI scan parameters for a chosen set of regions-of-interest (ROI). Reporting of data quality metrics (such as tSNR and image SNR (iSNR)) is often performed over the whole brain or for a small number of manually defined voxels. Importantly, this can obscure important inter-regional differences in data quality metrics. There is currently a lack of easy-to-use tools to analyse these data quality metrics, as well as functional statistical measures (such as beta maps or similar) simultaneously across multiple ROIs.
+
+The goal of the fMRI ROI Analysis Tool (fRAT), presented here, is to provide a toolset to address this gap, and provide a straightforward method for conducting ROI-wise analyses of fMRI metrics.
 
 # Statement of need
 
-`fRAT` is an open source, python-based analysis application, used to simplify the processing and analysis of fMRI data by converting voxelwise maps into ROI-wise maps (\autoref{fig:brain_images}). The graphical user interface of `fRAT` is designed to provide a user-friendly way to run and customise the settings of `fRAT` (\autoref{fig:GUI}). `fRAT` relies heavily on the python library `Nipype` to access the analysis tools provided by `FSL` [@jenkinsonFSL2012]. [Documentation](https://fmri-roi-analysis-tool.readthedocs.io/en/latest/) is available to provide information on installation and usage of `fRAT`.
[…]
