+++ paper/paper.bib
@@ -9,11 +9,105 @@
   pages = {930--933},
   issn = {1522-2594},
   doi = {10.1002/mrm.1124},
+  urldate = {2022-12-09},
   abstract = {The signal change in fMRI is dependent on the echo time and the rate of decay of transverse magnetization. The latter factor may vary across regions of the brain as a result of variations in field homogeneity. Previous measurements of the signal change with echo time have generally employed relatively slow multi-echo techniques, which may be sensitive to movement and habituation effects. Here a fast T measurement technique, involving the generation of six low-resolution echo planar images from a single FID, is described, and its use in the evaluation of the optimum echo time for visual, motor, and auditory fMRI experiments at 3.0 T is outlined. Magn Reson Med 45:930\textendash 933, 2001. \textcopyright{} 2001 Wiley-Liss, Inc.},
   langid = {english},
   keywords = {echo time,EPI,fMRI,multiple gradient-echo,T},
-  annotation = {\_eprint: https://onlinelibrary.wiley.com/doi/pdf/10.1002/mrm.1124},
   file = {/Users/lpxeh10/Zotero/storage/RJHS4KB2/Clare et al. - 2001 - Single-shot T measurement to establish optimum ech.pdf;/Users/lpxeh10/Zotero/storage/NTA7III5/mrm.html}
+}
+
+@article{estebanFMRIPrepRobustPreprocessing2019a,
+  title = {{{fMRIPrep}}: A Robust Preprocessing Pipeline for Functional {{MRI}}},
+  shorttitle = {{{fMRIPrep}}},
+  author = {Esteban, Oscar and Markiewicz, Christopher J. and Blair, Ross W. and Moodie, Craig A. and Isik, A. Ilkay and Erramuzpe, Asier and Kent, James D. and Goncalves, Mathias and DuPre, Elizabeth and Snyder, Madeleine and Oya, Hiroyuki and Ghosh, Satrajit S. and Wright, Jessey and Durnez, Joke and Poldrack, Russell A. and Gorgolewski, Krzysztof J.},
+  year = {2019},
+  month = jan,
+  journal = {Nature Methods},
+  volume = {16},
+  number = {1},
+  pages = {111--116},
+  publisher = {{Nature Publishing Group}},
+  issn = {1548-7105},
+  doi = {10.1038/s41592-018-0235-4},
+  urldate = {2023-04-25},
+  abstract = {Preprocessing of functional magnetic resonance imaging (fMRI) involves numerous steps to clean and standardize the data before statistical analysis. Generally, researchers create ad hoc preprocessing workflows for each dataset, building upon a large inventory of available tools. The complexity of these workflows has snowballed with rapid advances in acquisition and processing. We introduce fMRIPrep, an analysis-agnostic tool that addresses the challenge of robust and reproducible preprocessing for fMRI data. fMRIPrep automatically adapts a best-in-breed workflow to the idiosyncrasies of virtually any dataset, ensuring high-quality preprocessing without manual intervention. By introducing visual assessment checkpoints into an iterative integration framework for software testing, we show that fMRIPrep robustly produces high-quality results on a diverse fMRI data collection. Additionally, fMRIPrep introduces less uncontrolled spatial smoothness than observed with commonly used preprocessing tools. fMRIPrep equips neuroscientists with an easy-to-use and transparent preprocessing workflow, which can help ensure the validity of inference and the interpretability of results.},
+  copyright = {2018 The Author(s), under exclusive licence to Springer Nature America, Inc.},
+  langid = {english},
+  keywords = {Computational neuroscience,Image processing,Magnetic resonance imaging,Software,Standards},
[…]
