### 2023-04-26 #5200 @elliohow
Hi @mstimberg, here is the revised version of the paper. I'm not quite sure where the error is occuring as the pdf seems to have generated okay.

Tagging my co-authors: @schluppeck & @ppzstf
