### 2023-04-26 elliohow/fMRI_ROI_Analysis_Tool#37 @elliohow
Functionally derived ROIs can be problematic in many circumstances as this can result in a circular assessment of the data (www.ncbi.nlm.nih.gov/pubmed/20571517), as described by Geuter et al. 2018 (https://www.biorxiv.org/content/10.1101/295048v1.full). Instead it is often preferable to use  atlas derived ROIs based on anatomy, previous literature, or a separate localizer scan when a strong a priori hypothesis exists. We have now modified the manuscript to make this point.
