### 2023-04-26 elliohow/fMRI_ROI_Analysis_Tool#36 @elliohow
> The JOSS paper includes a statement of need, but there is none included in the documentation. The version from the paper should be included (and possibly expanded upon).

Apologies for this omission, we have now added the statement of need to the documentation page.

> For both, I think a fuller description of what scientific approaches are enabled or facilitated by this package (with citations, if they already exist) would be good. Or what problems it helps avoid. Basically, it's not clear to me why a user would want to use ROI analysis over the standard voxelwise approach.

Thank you for this comment.  We agree that this case should be made more clearly in the paper and have now modified the manuscript (within the constraints of the word limit) to make this point.

For the benefit of the review process, our thinking is explained here briefly with two particular use-case scenarios (which will be added into the documentation), one for planning of studies and one for data interrogation from studies:

1)	Study planning: When designing an fMRI study there is a trade-off between multiband (MB) factor and in-plane acceleration (SENSE, GRAPPA, ASSET) in the EPI acquisition. Multiband excites multiple imaging slices at the same time leading to significant scan time reduction, whilst in-plane acceleration allows reduction in geometric distortions and blurring due to T2* relaxation. Noise amplification tends to be non-uniform over the image (as it depends on coil sensitivity profiles), resulting in non-uniform temporal signal-to-noise ratio (tSNR) for fMRI. Therefore, it is beneficial to perform in-plane and MB acceleration at the same time, but it is important to find a good trade-off between artifact reduction and faster sampling. Also, different areas of the brain have different T2* (for example, short T2* in areas such as the amygdala and longer T2* in areas such as parietal cortex).  fRAT provides a basis in which to test fMRI data sets collected to determine the optimal multiband, in-plane acceleration, and echo time of choice for a study by assessing the tSNR across brain regions of interest to a study. An ROI analysis of tSNR can highlight inter-regional differences and facilitate the choice of MRI parameters for a study.

2)	Data interrogation: Voxel-wise maps are widely used and can provide useful information about how image, tSNR and other quality metrics vary across the brain. However, many plans for studies start out with strong assumptions about regions where functional responses are expected, but no clear information about statistical power, the participant numbers and scan duration required to detect effects of a particular size. To perform statistical power calculations, a commonly used approach is to consider an effect size observed in a particular region of interest and use appropriate univariate methods to calculate the statistical power for a given number of observations (or vice versa). While ROIs can be functionally derived, such functionally derived ROIs can overestimate effect sizes (https://www.biorxiv.org/content/10.1101/295048v1.full), so using ROIs derived from an atlas (as in the case of fRAT) is preferable.
