### 2023-04-26 #5200 @mstimberg
Thank you for the updated manuscript @elliohow , I think the error is some GitHub API hiccup that we can ignore.

@ZeitgeberH and @billbrod , please have a look at the updated paper and check whether it addresses your concerns :pray:
