### 2023-04-26 elliohow/fMRI_ROI_Analysis_Tool#37 @billbrod
Excellent. And it looks like this is included at the end of the [statement of need](https://fmri-roi-analysis-tool.readthedocs.io/en/latest/statement_of_need.html), right? I'd highlight it more: circularity and inflated effect size is a big problem of fMRI, so I'd really underline the importance of avoiding this -- pull the related text out of that last paragraph, maybe make it its own section. Right now, it sounds like an afterthought, because it's halfway through the last paragraph, which starts with "an example use-case". What do you think?
