### 2023-04-26 elliohow/fMRI_ROI_Analysis_Tool#37 @elliohow
Yes that makes sense. I will highlight it more in the documentation's statement of need. Is it ok as it currently is in the paper?
