### 2023-04-26 elliohow/fMRI_ROI_Analysis_Tool#37 @billbrod
Yup! Once you make those changes in the docs version, I'll close this issue.
