### 2023-04-26 elliohow/fMRI_ROI_Analysis_Tool#37 @elliohow
I've updated the statement of need page of the documentation.
