### 2023-04-28 #5200 @elliohow
> This paper is much clearer! My final concern is that it's not clear to me from the paper (or statement of need in the docs) that the ROIs `fRAT` uses are derived from anatomical atlases that the package provides (so I disagree with @ZeitgeberH that this is clear right now) -- I realize that from the conversations we've had during the review process, and it's implicit in the paper. I think it just needs an extra sentence, e.g., in the third paragraph of the paper:
>
> > The user provides a 4D fMRI timeseries as an input, from which the voxelwise maps of data quality metrics (e.g. tSNR) are computed within fRAT. Alternatively, pre-computed data quality maps can be used. In addition, the ROI analysis requires a structural scan (MPRAGE) which should be skull stripped (using FSL’s BET (Smith, 2002), or preferably, using optiBET (Lutkenhoff et al., 2014)) in order to run (Figure 3).
>
> Then add: "`fRAT` matches [or analyzes or something] the structural scan with the user's choice of one of more than a dozen ready-to-use anatomical atlases to define ROIs. These ROIs are used ..." and then continue that paragraph. Currently, I don't think it's clear how ROIs are defined, and the use of the anatomical atlases is important.
>
> Once that change is made (in both paper and docs), I recommend acceptance.

Instead of changing the 3rd paragraph, what about changing the 2nd paragraph from:
> `fRAT` relies heavily on the python library `Nipype` to access the analysis tools provided by FSL [@jenkinsonFSL2012].

To:
> `fRAT` relies heavily on the analysis tools and ROI atlases provided by FSL [@jenkinsonFSL2012].
