+++ paper/paper.md
@@ -34,13 +34,13 @@ The goal of the fMRI ROI Analysis Tool (fRAT), presented here, is to provide a t
 
 # Statement of need
 
-There are several tools that provide functionality to report ROI-wise summaries, including the widely used `Freesurfer` infrastructure [@fischlFreeSurfer2012] and packages built on top of them. For example, the `R` packages `ggseg` and `ggseg3d` [@mowinckelVisualizationBrainStatistics2020] can be used to show aggregated data such as cortical thickness in atlas-derived regions of interest. However these packages are designed primarily for use with anatomical datasets and would require some additional coding for use with fMRI data quality and statistical metrics. Several tools do provide data quality metrics for fMRI datasets, such as `fMRIPrep` [@estebanFMRIPrepRobustPreprocessing2019a] and `MRIQC` [@estebanMRIQCAdvancingAutomatic2017]. However, these tools either report voxelwise maps or aggregate metrics over the entire brain instead of chosen ROIs. This can obscure important inter-regional differences which may be particularly informative for optimizing scanning parameters for planned experiments
+There are several tools that provide functionality to report ROI-wise summaries, including the widely used `Freesurfer` infrastructure [@fischlFreeSurfer2012] and packages built on top of them. For example, the `R` packages `ggseg` and `ggseg3d` [@mowinckelVisualizationBrainStatistics2020] can be used to show aggregated data such as cortical thickness in atlas-derived regions of interest. However these packages are designed primarily for use with anatomical datasets and would require some additional coding for use with fMRI data quality and statistical metrics. Several tools do provide data quality metrics for fMRI datasets, such as `fMRIPrep` [@estebanFMRIPrepRobustPreprocessing2019a] and `MRIQC` [@estebanMRIQCAdvancingAutomatic2017]. However, these tools either report voxelwise maps or aggregate metrics over the entire brain instead of chosen ROIs. This can obscure important inter-regional differences which may be particularly informative for optimizing scanning parameters for planned experiments.
 
-`fRAT` is an open source, python-based application which focuses on ROI-wise analysis of fMRI data, by providing an easy to use and flexible pipeline for converting voxelwise data into ROI-wise data. The provided plotting options can be used to customize different aspects of the data, such as the spatial distribution of the metric of interest (\autoref{fig:brain_images}), while the statistical tools facilitate univariate and multivariate analyses within ROIs. The graphical user interface is designed to provide a user-friendly way to run and customize the settings of `fRAT` (\autoref{fig:GUI}) and [documentation](https://fmri-roi-analysis-tool.readthedocs.io/en/latest/) is available to provide information on installation and usage of fRAT. `fRAT` relies heavily on the python library `Nipype` to access the analysis tools provided by FSL [@jenkinsonFSL2012].
+`fRAT` is an open source, python-based application which focuses on ROI-wise analysis of fMRI data, by providing an easy to use and flexible pipeline for converting voxelwise data into ROI-wise data, based on `FSL`'s [@jenkinsonFSL2012] provided anatomical atlases. The provided plotting options can be used to customize different aspects of the data, such as the spatial distribution of the metric of interest (\autoref{fig:brain_images}), while the statistical tools facilitate univariate and multivariate analyses within ROIs. The graphical user interface is designed to provide a user-friendly way to run and customize the settings of `fRAT` (\autoref{fig:GUI}) and [documentation](https://fmri-roi-analysis-tool.readthedocs.io/en/latest/) is available to provide information on installation and usage of fRAT. `fRAT` relies heavily on the python library `Nipype` to access the analysis tools provided by `FSL` [@jenkinsonFSL2012].
 
[…]
