+++ docs/statement_of_need.rst
@@ -10,7 +10,7 @@ Overview
 
 There are several tools that provide functionality to report ROI-wise summaries, including the widely used `Freesurfer` infrastructure and packages built on top of them. For example, the `R` packages `ggseg` and `ggseg3d` can be used to show aggregated data such as cortical thickness in atlas-derived regions of interest. However these packages are designed primarily for use with anatomical datasets and would require some additional coding for use with fMRI data quality and statistical metrics. Several tools do provide data quality metrics for fMRI datasets, such as `fMRIPrep` and `MRIQC`. However, these tools either report voxelwise maps or aggregate metrics over the entire brain instead of chosen ROIs. This can obscure important inter-regional differences which may be particularly informative for optimizing scanning parameters for planned experiments
 
-`fRAT` is an open source, python-based application which focuses on ROI-wise analysis of fMRI data, by providing an easy to use and flexible pipeline for converting voxelwise data into ROI-wise data. The provided plotting options can be used to customize different aspects of the data, such as the spatial distribution of the metric of interest, while the statistical tools facilitate univariate and multivariate analyses within ROIs. The graphical user interface is designed to provide a user-friendly way to run and customize the settings of `fRAT`. `fRAT` relies heavily on the python library `Nipype` to access the analysis tools provided by FSL.
+`fRAT` is an open source, python-based application which focuses on ROI-wise analysis of fMRI data, by providing an easy to use and flexible pipeline for converting voxelwise data into ROI-wise data, based on `FSL`'s provided anatomical atlases. The provided plotting options can be used to customize different aspects of the data, such as the spatial distribution of the metric of interest, while the statistical tools facilitate univariate and multivariate analyses within ROIs. The graphical user interface is designed to provide a user-friendly way to run and customize the settings of `fRAT`. `fRAT` relies heavily on the python library `Nipype` to access the analysis tools provided by FSL.
 
 Usage
 -----
