### 2023-05-03 f52c252
Minor formatting fixes in paper.md
