### 2023-05-03 56508f8
Fix capitalization in paper.bib
