+++ paper/paper.bib
@@ -93,7 +93,7 @@
 }
 
 @misc{geuterEffectSizePower2018,
-  title = {Effect {{Size}} and {{Power}} in {{fMRI Group Analysis}}},
+  title = {Effect Size and Power in {{fMRI}} Group Analysis},
   author = {Geuter, Stephan and Qi, Guanghao and Welsh, Robert C. and Wager, Tor D. and Lindquist, Martin A.},
   year = {2018},
   month = apr,
@@ -148,7 +148,7 @@
 }
 
 @article{lutkenhoffOptimizedBrainExtraction2014,
-  title = {Optimized {{Brain Extraction}} for {{Pathological Brains}} ({{optiBET}})},
+  title = {Optimized Brain Extraction for Pathological Brains ({{optiBET}})},
   author = {Lutkenhoff, Evan S. and Rosenberg, Matthew and Chiang, Jeffrey and Zhang, Kunyu and Pickard, John D. and Owen, Adrian M. and Monti, Martin M.},
   year = {2014},
   month = dec,
@@ -167,7 +167,7 @@
 }
 
 @article{mowinckelVisualizationBrainStatistics2020,
-  title = {Visualization of {{Brain Statistics With R Packages}} Ggseg and Ggseg3d},
+  title = {Visualization of Brain Statistics With R Packages {ggseg} and {ggseg3d}},
   author = {Mowinckel, Athanasia M. and {Vidal-Pi{\~n}eiro}, Didac},
   year = {2020},
   month = dec,
@@ -256,7 +256,7 @@
 }
 
 @article{welvaertDefinitionSignalToNoiseRatio2013,
-  title = {On the {{Definition}} of {{Signal-To-Noise Ratio}} and {{Contrast-To-Noise Ratio}} for {{fMRI Data}}},
+  title = {On the Definition of Signal-To-Noise Ratio and Contrast-To-Noise Ratio for {{fMRI}} Data},
   author = {Welvaert, Marijke and Rosseel, Yves},
   year = {2013},
   month = nov,
