### 2023-05-03 eb2fbe7
* Replaced citation * Removed unnecessary acronyms
