+++ paper/paper.bib
@@ -57,24 +57,6 @@
   file = {/Users/lpxeh10/Zotero/storage/7MWG7P5V/Esteban et al. - 2017 - MRIQC Advancing the automatic prediction of image.pdf}
 }
 
-@article{fazalComparisonMultibandMultiband2022,
-  title = {A Comparison of Multiband and Multiband Multiecho Gradient-echo {{EPI}} for Task {{fMRI}} at 3~{{T}}},
-  author = {Fazal, Zahra and Gomez, Daniel E. P. and Llera, Alberto and Marques, Jos{\'e} P. R. F. and Beck, Thomas and Poser, Benedikt A. and Norris, David G.},
-  year = {2022},
-  month = oct,
-  journal = {Human Brain Mapping},
-  volume = {44},
-  number = {1},
-  pages = {82--93},
-  issn = {1065-9471},
-  doi = {10.1002/hbm.26081},
-  urldate = {2023-04-25},
-  abstract = {A multiband (MB) echo-planar imaging (EPI) sequence is compared to a multiband multiecho (MBME) EPI protocol to investigate differences in sensitivity for task functional magnetic resonance imaging (fMRI) at 3~T. Multiecho sampling improves sensitivity in areas where single-echo-EPI suffers from dropouts. However, It requires in-plane acceleration to reduce the echo train length, limiting the slice acceleration factor and the temporal and spatial resolution Data were acquired for both protocols in two sessions 24\,h apart using an adapted color-word interference Stroop task. Besides protocol comparison statistically, we performed test\textendash retest reliability across sessions for different protocols and denoising methods. We evaluated the sensitivity of two different echo-combination strategies for MBME-EPI. We examined the performance of three different data denoising approaches: ``Standard,'' ``AROMA,'' and ``FIX'' for MB and MBME, and assessed whether a specific method is preferable. We consider using an appropriate autoregressive model order within the general linear model framework to correct TR differences between the protocols. The comparison between protocols and denoising methods showed at group level significantly higher mean z-scores and the number of active voxels for MBME in the motor, subcortical and medial frontal cortices. When comparing different echo combinations, our results suggest that a contrast-to-noise ratio weighted echo combination improves sensitivity in MBME compared to simple echo-summation. This study indicates that MBME can be a preferred protocol in task fMRI at spatial resolution ({$\geq$}2\,mm), primarily in medial prefrontal and subcortical areas., In summary, we found that multiband multiecho (MBME) and MB give a comparable performance in most brain regions with this comparison study. In the regions with dropouts and susceptibility induced inhomogeneity, MBME performed somewhat better at the group level. The results showed that the difference between sequences is more significant than between different denoising (FIX and AROMA) strategies at the group level, highlighting the importance of considering an optimized sequence before data collection. For MBME, choosing an optimal echo combination can influence the sensitivity at the group-level analysis.},
-  pmcid = {PMC9783458},
-  pmid = {36196782},
-  file = {/Users/lpxeh10/Zotero/storage/2VWWBHQI/Fazal et al. - 2022 - A comparison of multiband and multiband multiecho .pdf}
-}
-
 @article{fischlFreeSurfer2012,
   title = {{{FreeSurfer}}},
   author = {Fischl, Bruce},
