### 2023-05-04 50c2115
* Began separation of utilities from map creation
