+++ fRAT/utils/statmap.py
@@ -5,6 +5,7 @@ from nipype.interfaces import fsl
 import logging
 import time
 
+from ..HOUSE.handler import HOUSE
 from .utils import *
 
 config = None
@@ -12,7 +13,7 @@ config_path = ''
 config_filename = ''
 
 
-def file_setup(func):
+def statmap_file_setup(func):
     file_location = config.input_folder_name
 
     if config.output_folder_name != 'DEFAULT':
@@ -21,8 +22,6 @@ def file_setup(func):
         output_folder = 'statmaps/imageSNR_report'
     elif func == 'Temporal SNR':
         output_folder = 'statmaps/temporalSNR_report'
-    elif func == 'Add Gaussian noise':
-        output_folder = 'added_noise'
 
     if config.base_folder in ("", " "):
         print('Select the directory which contains the subject folders.')
@@ -39,14 +38,11 @@ def file_setup(func):
         # Find all nifti and analyze files
         participants[participant] = Utils.find_files(f"{participant}/{file_location}", "hdr", "nii.gz", "nii")
 
-        if 'statmaps' in output_folder:
-            Utils.check_and_make_dir(f"{participant}/statmaps")
-            Utils.check_and_make_dir(f"{participant}/{output_folder}", delete_old=True)
-            Utils.save_config(f"{participant}/{output_folder}", config_path, config_filename,
-                              additional_info=[f"statistical_map_created = '{func}'\n"],
-                              new_config_name='statmap_config')
-        else:
-            Utils.check_and_make_dir(f"{participant}/{file_location}/{output_folder}", delete_old=True)
+        Utils.check_and_make_dir(f"{participant}/statmaps")
+        Utils.check_and_make_dir(f"{participant}/{output_folder}", delete_old=True)
+        Utils.save_config(f"{participant}/{output_folder}", config_path, config_filename,
+                          additional_info=[f"statistical_map_created = '{func}'\n"],
+                          new_config_name='statmap_config')
 
     return participants, output_folder, file_location
 
@@ -57,13 +53,6 @@ def calculate_sigma_in_volumes(file_path):
 
     # Equation found here: https://www.jiscmail.ac.uk/cgi-bin/webadmin?A2=FSL;f6fd75a6.1709
     return 1 / (2 * config.highpass_filter_cutoff * TR)
-
-
-def save_brain(data, ext, no_ext_file, output_folder, header=None):
-    brain = nib.Nifti1Pair(data, None, header)
-    nib.save(brain, f"{output_folder}/{no_ext_file}{ext}.nii.gz")
-
-    return f"{output_folder}/{no_ext_file}{ext}.nii.gz"
 
 
 def temporalSNR_calc(file, no_ext_file, output_folder):
@@ -125,7 +114,7 @@ def separate_noise_from_func(file, no_ext_file, output_folder, participant):
     else:
         raise Exception('Noise volume location not valid.')
 
-    noise_file = save_brain(noise_data, '_noise_volume', no_ext_file, output_folder, header)
+    noise_file = Utils.save_brain(noise_data, '_noise_volume', no_ext_file, output_folder, header)
 
     func_file = save_to_cleaned_folder(func_data, header, no_ext_file, participant,
                                        f'Removed noise volume from {config.noise_volume_location.lower()} of timeseries')
@@ -138,7 +127,7 @@ def save_to_cleaned_folder(data, header, no_ext_file, participant, method):
         f.seek(0)  # Go to start of file
         f.write(f'{no_ext_file}: {method}\n')
 
-    data = save_brain(data, '', no_ext_file, f'{participant}/{config.input_folder_name}_cleaned', header)
[…]
