+++ fRAT/HOUSE/__init__.py
@@ -1,0 +1,2 @@
+from .add_motion import *
+from .add_noise import *
