+++ fRAT/HOUSE/handler.py
@@ -1,22 +1,27 @@
 import itertools
 import os
+import importlib
 from pathlib import Path
 
-from . import *
 from ..utils import Utils
 
 config = None
+UTILITIES = None
+
+def import_utilities():
+    global UTILITIES
+    from . import __all__
+
+    UTILITIES = [importlib.import_module(f".{utility}", package=__package__) for utility in __all__]
 
 
-def file_setup(func):
+def file_setup(utility):
     file_location = config.input_folder_name
 
     if config.output_folder_name != 'DEFAULT':
-        output_folder = f"statmaps/{config.output_folder_name}"
-    elif func == 'Add Gaussian noise':
-        output_folder = 'added_noise'
-    elif func == 'Add motion':
-        output_folder = 'added_motion'
+        output_folder = config.output_folder_name
+    else:
+        output_folder = utility.FOLDER_NAME
 
     if config.base_folder in ("", " "):
         print('Select the directory which contains the subject folders.')
@@ -34,60 +39,90 @@ def file_setup(func):
         participants[participant] = Utils.find_files(f"{participant}/{file_location}", "hdr", "nii.gz", "nii")
 
         # Make directory to save results
-        Utils.check_and_make_dir(f"{participant}/{file_location}/{output_folder}", delete_old=True)
+        if isinstance(output_folder, list):
+            for folder in output_folder:
+                Utils.check_and_make_dir(f"{participant}/{folder}", delete_old=True)
+        else:
+            Utils.check_and_make_dir(f"{participant}/{output_folder}", delete_old=True)
 
     return participants, output_folder, file_location
 
 
-def utility_handler(participants, output_folder, file_location, func):
+def utility_handler(participants, output_folder, file_location, utility):
     if config.multicore_processing:
         pool = Utils.start_processing_pool()
     else:
         pool = None
 
-    if config.verbose:
-        print(f'\nSaving output in directory: {config.input_folder_name}/{output_folder}')
+    if config.verbose and isinstance(output_folder, list):
+        direcs = ', '.join(output_folder)
+        print(f'\nSaving output in directories: {direcs}')
+    elif config.verbose and len(output_folder) == 1:
+        print(f'\nSaving output in directory: {output_folder}')
 
-    for participant, files in participants.items():
-        participant_dir = f"{participant}/{file_location}"
-
+    for participant_dir, files in participants.items():
         if config.verbose:
-            print(f"\nRunning '{func}' utility on participant: {participant_dir.split('/')[-2]}"
+            print(f"\nRunning '{utility.UTILITY_NAME}' utility on participant: {participant_dir.split('/')[-1]}"
                   f"\n      Running utility on {len(files)} files")
 
         iterable = zip(files,
                        itertools.repeat(participant_dir),
+                       itertools.repeat(file_location),
                        itertools.repeat(output_folder),
-                       itertools.repeat(func),
+                       itertools.repeat(utility),
                        itertools.repeat(config))
 
         if config.multicore_processing:
-            list(pool.starmap(run_utility, iterable))
+            return_val = list(pool.starmap(run_utility, iterable))
 
         else:
-            list(itertools.starmap(run_utility, iterable))
+            return_val = list(itertools.starmap(run_utility, iterable))
+
+        if return_val[0] is not None:
+                iterable = zip(files,
+                               itertools.repeat(participant_dir),
+                               itertools.repeat(file_location),
+                               itertools.repeat(output_folder),
+                               itertools.repeat(utility),
+                               itertools.repeat(config),
+                               itertools.repeat(return_val))
+
+                if config.multicore_processing:
[…]
