+++ fRAT/HOUSE/__init__.py
@@ -1,2 +1,6 @@
-from .add_motion import *
-from .add_noise import *
+from os.path import dirname, basename, isfile, join
+import glob
+
+modules = glob.glob(join(dirname(__file__), "*.py"))
+
+__all__ = [basename(f)[:-3] for f in modules if isfile(f) and not f.endswith(("handler.py", "__init__.py"))]
