### 2023-05-11 #5200 @elliohow
Thank you @billbrod, @ZeitgeberH & @mstimberg for being so thorough in your reviews, the software and paper has benefited immensely from the process, especially in terms of making `fRAT` multi platform.
